+++ core/src/mapache/defaults.rs
@@ -7,8 +7,8 @@ use crate::{commands::Compression, repository::repo::RepoConfig, utils::size};
 pub(crate) const APP_NAME: &str = "mapache";
 
 // --- Concurrency ---
-pub(crate) const DEFAULT_READ_CONCURRENCY: usize = 2;
-pub(crate) const DEFAULT_WRITE_CONCURRENCY: usize = 4;
+pub(crate) const DEFAULT_SNAPSHOT_READERS: usize = 2;
+pub(crate) const DEFAULT_SNAPSHOT_PACKERS: usize = 4;
 
 // --- Index ---
 pub(crate) const INDEX_FLUSH_TIMEOUT: Duration = Duration::from_secs(10 * 60);