+++ core/src/repository/gc.rs
@@ -165,7 +165,7 @@ impl Plan {
         // No need to repack and rewrite the indices if there are no obsolete packs
         if !self.obsolete_packs.is_empty() {
             self.repo
-                .init_pack_saver(mapache::defaults::DEFAULT_WRITE_CONCURRENCY)?;
+                .init_pack_saver(mapache::defaults::DEFAULT_SNAPSHOT_PACKERS)?;
 
             self.repack()?;
             let repo_stats = self.repo.flush_and_finalize_pack_saver()?;