+++ core/tests/integration_tests/test_cmd_mount.rs
@@ -174,8 +174,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: true,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -308,8 +308,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: true,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -333,8 +333,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: true,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)