+++ core/tests/integration_tests/test_cmd_restore.rs
@@ -82,8 +82,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: true,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -207,8 +207,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: true,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -291,8 +291,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: true,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -407,8 +407,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: true,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -556,8 +556,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: true,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -701,8 +701,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: true,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -844,8 +844,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: true,
             parent: UseSnapshot::Latest,
-            read_concurrency: 1,
-            write_concurrency: 1,
+            num_readers: 1,
+            num_packers: 1,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)