+++ core/tests/integration_tests/test_cmd_snapshot.rs
@@ -81,8 +81,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: false,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -204,8 +204,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: false,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: true,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -305,8 +305,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: false,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -435,8 +435,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: false,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -458,8 +458,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: false,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -575,8 +575,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: false,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -689,8 +689,8 @@ mod tests {
             skip_if_unchanged: false,
             no_scan: false,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };
         commands::cmd_snapshot::run(&global, &snapshot_args)
@@ -809,8 +809,8 @@ mod tests {
             skip_if_unchanged: true,
             no_scan: false,
             parent: UseSnapshot::Latest,
-            read_concurrency: 2,
-            write_concurrency: 2,
+            num_readers: 2,
+            num_packers: 2,
             dry_run: false,
         };